+++ src/Task2/task2.h
@@ -1,9 +1,9 @@
 #ifndef TASK2_H
 #define TASK2_H
 
-extern double screen[4];
-extern unsigned long int TASK2_MAX_ITER;
+extern long double screen[4];
+extern unsigned long long int TASK2_MAX_ITER;
 
-void TASK2_run(unsigned long int  num_points, unsigned long int  batch_size, int  num_threads);
+void TASK2_run(unsigned long long int num_points, unsigned long long int batch_size, int num_threads);
 
 #endif
